Lab 3/requirements.txt
Adafruit-Blinka==6.2.2
Adafruit-PlatformDetect==3.1.1
Adafruit-PureIO==1.1.8
adafruit-circuitpython-apds9960==2.2.5
adafruit-circuitpython-busdevice==5.0.5
adafruit-circuitpython-register==1.9.4
adafruit-circuitpython-rgb-display==3.10.5
circuitpython-i2c-button==2.2.1
colour==0.1.5
Pillow==9.0.1
pyftdi==0.52.9
pyserial==3.5
pyusb==1.1.1
rpi-ws281x==4.2.5
RPI.GPIO==0.7.1
spidev==3.5
sysv-ipc==1.1.0
webcolors==1.11.1
sparkfun-qwiic-i2c==0.9.11
sparkfun-qwiic-button==2.0.1
